backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1

SQLAlchemy==2.0.36
psycopg[binary]==3.2.3

python-jose==3.3.0

passlib==1.7.4
bcrypt==4.1.3

cryptography==43.0.3

pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.2.1
email-validator==2.2.0
python-multipart==0.0.12

webauthn==2.5.0
cbor2==5.8.0
pyOpenSSL==25.1.0
asn1crypto==1.5.1

jinja2==3.1.4
itsdangerous==2.2.0

aiofiles==24.1.0
requests==2.32.3

scikit-learn==1.5.2
numpy==2.1.3
scipy==1.17.1
joblib==1.5.3
threadpoolctl==3.6.0